🚶‍♀️ refactor to a walkPaths utility
